refactor(connectors)!: move connector routes to /v1alpha/admin/connectors
Connectors are admin controlled resources.

Connector routes previously served at /v1beta/connectors/* are now served
at /v1alpha/admin/connectors/* as part of the admin router.

Admin protocol now extends Connectors, so AdminImpl satisfies both
protocols. AdminImpl delegates connector methods to the injected
ConnectorServiceImpl. The standalone connectors FastAPI router is removed;
admin/fastapi_routes.py owns all connector routes.
