+++ docs/docs/api-overview.mdx
@@ -34,7 +34,7 @@ OGX implements the OpenAI API and organizes endpoints by stability level. Use an
         { title: 'Rerank', endpoint: '/v1alpha/inference/rerank', href: '/docs/api-experimental/rerank-v-1-alpha-inference-rerank-post', blurb: 'Document reranking for search relevance' },
         { title: 'File Processors', endpoint: '/v1alpha/file_processors', href: '/docs/api-experimental/file-processors', blurb: 'Document ingestion and chunking' },
         { title: 'Interactions', endpoint: '/v1alpha/interactions', href: '/docs/api-experimental/interactions', blurb: 'Google Interactions API compatibility layer' },
-        { title: 'Connectors', endpoint: '/v1beta/connectors', href: '/docs/api-experimental/connectors', blurb: 'External tool and service connectors' },
+        { title: 'Connectors', endpoint: '/v1alpha/admin/connectors', href: '/docs/api-experimental/connectors', blurb: 'External tool and service connectors' },
       ],
     },
     {
