+++ docs/static/openai-coverage.json
@@ -1595,8 +1595,7 @@
                 {
                   "property": "POST.responses.200.content.application/json.properties.text",
                   "details": [
-                    "Type added: ['object']",
-                    "Default changed: None -> {'format': {'type': 'text'}}"
+                    "Type added: ['object']"
                   ]
                 },
                 {
