+++ src/ogx_api/admin/fastapi_routes.py
@@ -13,9 +13,18 @@ all API-related code together.
 
 from typing import Annotated
 
-from fastapi import APIRouter, Depends
+from fastapi import APIRouter, Depends, Path, Query
 
+from ogx_api.connectors.models import (
+    Connector,
+    GetConnectorRequest,
+    GetConnectorToolRequest,
+    ListConnectorsResponse,
+    ListConnectorToolsRequest,
+    ListToolsResponse,
+)
 from ogx_api.router_utils import create_path_dependency, create_query_dependency, standard_responses
+from ogx_api.tools import ToolDef
 from ogx_api.version import OGX_API_V1ALPHA
 
 from .api import Admin
@@ -32,6 +41,8 @@ from .models import (
 # Automatically generate dependency functions from Pydantic models
 get_inspect_provider_request = create_path_dependency(InspectProviderRequest)
 get_list_routes_request = create_query_dependency(ListRoutesRequest)
+get_connector_request = create_path_dependency(GetConnectorRequest)
+list_connector_tools_request = create_path_dependency(ListConnectorToolsRequest)
 
 
 def create_router(impl: Admin) -> APIRouter:
@@ -113,5 +124,52 @@ def create_router(impl: Admin) -> APIRouter:
     )
     async def version() -> VersionInfo:
         return await impl.version()
+
+    @router.get(
+        "/admin/connectors",
+        response_model=ListConnectorsResponse,
+        summary="List all connectors.",
+        description="List all configured connectors.",
+    )
+    async def list_connectors() -> ListConnectorsResponse:
+        return await impl.list_connectors()
+
+    @router.get(
+        "/admin/connectors/{connector_id}/tools/{tool_name}",
+        response_model=ToolDef,
+        summary="Get a tool by name from a connector.",
+        description="Get a tool definition by its name from a connector.",
+    )
+    async def get_connector_tool(
+        connector_id: Annotated[str, Path(description="Identifier for the connector")],
+        tool_name: Annotated[str, Path(description="Name of the tool")],
+        authorization: Annotated[str | None, Query(description="Authorization token")] = None,
+    ) -> ToolDef:
+        request = GetConnectorToolRequest(connector_id=connector_id, tool_name=tool_name)
+        return await impl.get_connector_tool(request, authorization=authorization)
+
+    @router.get(
+        "/admin/connectors/{connector_id}/tools",
+        response_model=ListToolsResponse,
+        summary="List tools from a connector.",
+        description="List all tools available from a connector.",
+    )
+    async def list_connector_tools(
+        request: Annotated[ListConnectorToolsRequest, Depends(list_connector_tools_request)],
+        authorization: Annotated[str | None, Query(description="Authorization token")] = None,
+    ) -> ListToolsResponse:
+        return await impl.list_connector_tools(request, authorization=authorization)
+
+    @router.get(
+        "/admin/connectors/{connector_id}",
+        response_model=Connector,
+        summary="Get a connector by its ID.",
+        description="Get a connector by its ID.",
+    )
+    async def get_connector(
+        request: Annotated[GetConnectorRequest, Depends(get_connector_request)],
+        authorization: Annotated[str | None, Query(description="Authorization token")] = None,
+    ) -> Connector:
+        return await impl.get_connector(request, authorization=authorization)
 
     return router
