+++ tests/unit/core/routers/test_connectors_router.py
@@ -28,7 +28,7 @@ from ogx_api import (
     ListToolsResponse,
     ToolDef,
 )
-from ogx_api.connectors.fastapi_routes import create_router
+from ogx_api.admin.fastapi_routes import create_router
 from ogx_api.connectors.models import (
     GetConnectorRequest,
     ListConnectorToolsRequest,
@@ -74,7 +74,7 @@ async def test_list_connectors_returns_empty_list():
     router = create_router(impl)
     app.include_router(router)
 
-    list_endpoint = _get_endpoint(router, "/v1beta/connectors", "GET")
+    list_endpoint = _get_endpoint(router, "/v1alpha/admin/connectors", "GET")
     response = await list_endpoint()
 
     assert response.data == []
@@ -91,7 +91,7 @@ async def test_list_connectors_returns_connectors():
     router = create_router(impl)
     app.include_router(router)
 
-    list_endpoint = _get_endpoint(router, "/v1beta/connectors", "GET")
+    list_endpoint = _get_endpoint(router, "/v1alpha/admin/connectors", "GET")
     response = await list_endpoint()
 
     assert len(response.data) == 1
@@ -113,7 +113,7 @@ async def test_get_connector_returns_connector():
     router = create_router(impl)
     app.include_router(router)
 
-    get_endpoint = _get_endpoint(router, "/v1beta/connectors/{connector_id}", "GET")
+    get_endpoint = _get_endpoint(router, "/v1alpha/admin/connectors/{connector_id}", "GET")
     request = GetConnectorRequest(connector_id="test-connector")
     response = await get_endpoint(request=request, authorization=None)
 
@@ -132,7 +132,7 @@ async def test_get_connector_with_authorization():
     router = create_router(impl)
     app.include_router(router)
 
-    get_endpoint = _get_endpoint(router, "/v1beta/connectors/{connector_id}", "GET")
+    get_endpoint = _get_endpoint(router, "/v1alpha/admin/connectors/{connector_id}", "GET")
     request = GetConnectorRequest(connector_id="test-connector")
     response = await get_endpoint(request=request, authorization="test-token")
 
@@ -151,7 +151,7 @@ async def test_get_connector_not_found_raises_error():
     router = create_router(impl)
     app.include_router(router)
 
-    get_endpoint = _get_endpoint(router, "/v1beta/connectors/{connector_id}", "GET")
+    get_endpoint = _get_endpoint(router, "/v1alpha/admin/connectors/{connector_id}", "GET")
     request = GetConnectorRequest(connector_id="nonexistent")
 
     with pytest.raises(ConnectorNotFoundError):
@@ -171,7 +171,7 @@ async def test_list_connector_tools_returns_tools():
     router = create_router(impl)
     app.include_router(router)
 
-    list_endpoint = _get_endpoint(router, "/v1beta/connectors/{connector_id}/tools", "GET")
+    list_endpoint = _get_endpoint(router, "/v1alpha/admin/connectors/{connector_id}/tools", "GET")
     request = ListConnectorToolsRequest(connector_id="test-connector")
     response = await list_endpoint(request=request, authorization=None)
 
@@ -189,7 +189,7 @@ async def test_list_connector_tools_empty():
     router = create_router(impl)
     app.include_router(router)
 
-    list_endpoint = _get_endpoint(router, "/v1beta/connectors/{connector_id}/tools", "GET")
+    list_endpoint = _get_endpoint(router, "/v1alpha/admin/connectors/{connector_id}/tools", "GET")
     request = ListConnectorToolsRequest(connector_id="test-connector")
     response = await list_endpoint(request=request, authorization=None)
 
@@ -206,7 +206,7 @@ async def test_list_connector_tools_with_authorization():
     router = create_router(impl)
     app.include_router(router)
 
-    list_endpoint = _get_endpoint(router, "/v1beta/connectors/{connector_id}/tools", "GET")
+    list_endpoint = _get_endpoint(router, "/v1alpha/admin/connectors/{connector_id}/tools", "GET")
     request = ListConnectorToolsRequest(connector_id="test-connector")
     _ = await list_endpoint(request=request, authorization="test-token")
 
@@ -227,7 +227,7 @@ async def test_get_connector_tool_returns_tool():
     router = create_router(impl)
     app.include_router(router)
 
-    get_endpoint = _get_endpoint(router, "/v1beta/connectors/{connector_id}/tools/{tool_name}", "GET")
+    get_endpoint = _get_endpoint(router, "/v1alpha/admin/connectors/{connector_id}/tools/{tool_name}", "GET")
     response = await get_endpoint(
         connector_id="test-connector",
         tool_name="test-tool",
@@ -249,7 +249,7 @@ async def test_get_connector_tool_with_authorization():
     router = create_router(impl)
     app.include_router(router)
 
-    get_endpoint = _get_endpoint(router, "/v1beta/connectors/{connector_id}/tools/{tool_name}", "GET")
+    get_endpoint = _get_endpoint(router, "/v1alpha/admin/connectors/{connector_id}/tools/{tool_name}", "GET")
     _ = await get_endpoint(
         connector_id="test-connector",
         tool_name="test-tool",
@@ -269,7 +269,7 @@ async def test_get_connector_tool_not_found_raises_error():
     router = create_router(impl)
     app.include_router(router)
 
-    get_endpoint = _get_endpoint(router, "/v1beta/connectors/{connector_id}/tools/{tool_name}", "GET")
+    get_endpoint = _get_endpoint(router, "/v1alpha/admin/connectors/{connector_id}/tools/{tool_name}", "GET")
 
     with pytest.raises(ConnectorToolNotFoundError):
         await get_endpoint(
@@ -292,10 +292,10 @@ def test_openapi_schema_has_connectors_endpoints():
     schema = app.openapi()
 
     # Verify all endpoints are documented
-    assert "/v1beta/connectors" in schema["paths"]
-    assert "/v1beta/connectors/{connector_id}" in schema["paths"]
-    assert "/v1beta/connectors/{connector_id}/tools" in schema["paths"]
-    assert "/v1beta/connectors/{connector_id}/tools/{tool_name}" in schema["paths"]
+    assert "/v1alpha/admin/connectors" in schema["paths"]
+    assert "/v1alpha/admin/connectors/{connector_id}" in schema["paths"]
+    assert "/v1alpha/admin/connectors/{connector_id}/tools" in schema["paths"]
+    assert "/v1alpha/admin/connectors/{connector_id}/tools/{tool_name}" in schema["paths"]
 
 
 def test_openapi_schema_list_connectors_is_get():
@@ -306,7 +306,7 @@ def test_openapi_schema_list_connectors_is_get():
     app.include_router(router)
 
     schema = app.openapi()
-    connectors_path = schema["paths"]["/v1beta/connectors"]
+    connectors_path = schema["paths"]["/v1alpha/admin/connectors"]
 
     assert "get" in connectors_path
     assert connectors_path["get"]["summary"] == "List all connectors."
@@ -320,7 +320,7 @@ def test_openapi_schema_get_connector_has_path_param():
     app.include_router(router)
 
     schema = app.openapi()
-    get_connector_path = schema["paths"]["/v1beta/connectors/{connector_id}"]
+    get_connector_path = schema["paths"]["/v1alpha/admin/connectors/{connector_id}"]
 
     assert "get" in get_connector_path
     parameters = get_connector_path["get"]["parameters"]
@@ -336,7 +336,7 @@ def test_openapi_schema_has_authorization_query_param():
     app.include_router(router)
 
     schema = app.openapi()
-    get_connector_path = schema["paths"]["/v1beta/connectors/{connector_id}"]
+    get_connector_path = schema["paths"]["/v1alpha/admin/connectors/{connector_id}"]
 
     parameters = get_connector_path["get"]["parameters"]
     auth_params = [p for p in parameters if p["name"] == "authorization"]
